rozwiązanie - rok urodzenia - int

+++ zadanie03.py
@@ -1,6 +1,6 @@
 ROK_BIEZACY = 2023
 
-rok_ur = float(input('Podaj rok urodzenia: '))
+rok_ur = int(input('Podaj rok urodzenia: '))
 wiek = ROK_BIEZACY - rok_ur
 print(wiek)
 if wiek >= 18:
